[#5] Handle slack response 429 Too Many Requests
Problem: Sometimes slack can respond with status 429 Too Many Requests.
We have to handle this and take a try to request the slack api again
after the delay set in Retry-After header.

Solution: Wrap enent handler with function which detects the response
with status code 429, extracts the delay from headers and runs again the
event handler after the delay. The number of attempts is defined in bot
configuration.