[#31] Allow users to DM the bot (#39)
Problem: Currently, if the user DMs the bot, this is treated as for
ordinary channels (reporting only errors to the sender if any). However,
such conversation can contain only bot and the user, and if the user DMs
the bot, probably he wants to get translation of his message himself.

Solution: Change handling of the DM message, sending translations of all
time references to the sender himself.