[#44] Add help command
Problem: Currently, each ephemeral with translations contain button
which functionality doubles existing entrypoint. Because it can be
annoying, we should remove it and instead give a hint to a user how to
call tzbot entrypoints.

Solution: Remove buttons from ephemerals, add command /tzhelp, append
the usage of that command randomly with rather low chance to ephemerals
with translations.