[#36] Express doubts in translations
Problem: In many cases we can't know the full date because we are not
able to get all the context of the time reference. In such situations we
should explain the user somehow that we are not 100% sure in the
inferred date.

Solution: Implement simple guessing logic; we are either totally sure or
sure only about time, output messages are changed correspondingly.