[#36] Add warnings about possible timeshifts around
Problem: In some cases the bot selects the date rather arbitrarily, and
if a timeshift is somewhere around, the translation may be dependent on
concrete date. We should warn a user about possible timeshifts.

Solution: For some types of date references, also compute a date range
to check for the possible timeshift, and report to user if something
found.