[#53] Handle matching timezones
Problem: Sometimes, when timezones in message and a receiver's one are the
same, it's not required to actually translate anything because the
original time reference is already relevant for the receiver.

Solution: for other users than sender, say "timezone is the same", for
sender himself translate only time references that are in other
timezones and ignore ones with his timezone, but translate everything
when modal is triggered.