[#68] Flush feedback to the file immediately
Problem: Right now, we're using `withFile` to open a file handle to
`feedback.log`. This seems to use `BlockBuffering` by default, which
means feedback is not immediately written to the log file. This might
take some time to be written to file or, worst case scenario, is only
written when the application shuts down.

Solution: Use line buffering instead, so that output is flushed to the
file as soon as a newline character is emitted.