[#107] Correctly parse `channel_join` events
Problem: The parser fails to parse events of type `message` and subtype
`channel_join`, and the logs are flooded with errors.

This seems to happen because the parser expects every event of type
`message` to have a `blocks` field, and that's not always true.

Solution: Make the `blocks` field optional.