[#109] Enhance `MessageContext`
Problem: The `MessageContext` data type adds a `message_id` label to log
messages, so we can identify which Slack message the server was
processing when the log event occurred.

However, a message ID (or message timestamp, really) is not enough to
identify a message. A message is uniquely identified by the `(timestamp,
channel_id)` pair.

Solution: Add a `channel_id` label to `MessageContext`.