[#104] Warn about sharing message with the dev team
Problem: When the user uses the "report an issue" feature, the message
with the time references will be sent to the dev team.

Seeing as this message could contain private information (e.g. it could
be in a private channel, or be in a DM, etc), we should warn the user
about this.

Solution: Warn the user that their message will be shared with the dev
team.