[Chore] Reorder keys of the json object used in `test_messageBlocksSpec`
Problem: The example JSON in `justEverythingRaw` appears to have been
aeson-decoded and then aeson-encoded, which means the object keys have
been reordered alphabetically (that's what aeson does).

However, this makes it harder to reconstruct this JSON if we ever need
to. It's simpler to just paste the JSON we get directly from Slack's Web
API.

Solution: Paste the JSON exactly as returned by Slack's Web API.